Versão inicial de uma calculadora simples criada em EJB 3.1

+++ calculadoraremotaClient/ejbModule/br/com/julianfernando/sessionbeans/CalculadoraIf.java
@@ -10,9 +10,9 @@ package br.com.julianfernando.sessionbeans;
  */
 public interface CalculadoraIf {
 
-	public int soma(int a, int b);
+	public double soma(double a, double b);
 	
-	public int subtracao(int a, int b);
+	public double subtracao(double a, double b);
 	
 	public double multiplicacao(double a, double b);
